requirements.txt
# Core Flask dependencies
Flask==2.3.3
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-Limiter==3.5.0

# Database dependencies
SQLAlchemy==2.0.23
psycopg2-binary==2.9.8

# Configuration and validation
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Security
bcrypt==4.1.2

# Development dependencies
pytest==7.4.3
pytest-flask==1.3.0
black==23.11.0
ruff==0.1.6
mypy==1.7.1

# Production server
gunicorn==21.2.0 